backend/requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.7
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.15.2
environ==1.0
psycopg2==2.9.10
sqlparse==0.5.3
typing_extensions==4.13.0
